requirements.txt
python-dotenv
pandas
numpy
requests
tensorflow
fastapi
mangum
uvicorn